+++ Scripts/Helper/RunEvent/DSC_Event_RunEventsByTrigger2D.cs
@@ -9,7 +9,16 @@ namespace DSC.Event.Helper
         #region Variable
 
         #region Variable - Inspector
-#pragma warning disable 0649
+
+#if UNITY_EDITOR
+
+        [TextField("Events Description", 3)]
+        [SerializeField] string m_sDescription;
+
+#endif
+
+        [TagField]
+        [SerializeField] protected string[] m_arrTargetTag;
 
         [Header("Condition")]
         [SerializeField] protected EventCondition[] m_arrCondition;
@@ -24,14 +33,13 @@ namespace DSC.Event.Helper
         [SerializeField] protected EventGameObject m_hStayEventGameObject;
         [SerializeField] protected EventGameObject m_hExitEventGameObject;
 
-#pragma warning restore 0649
         #endregion
 
         protected EventConditionData m_hConditionData;
 
         #endregion
 
-        #region Base - Mono
+        #region Unity
 
         protected virtual void Awake()
         {
@@ -40,6 +48,9 @@ namespace DSC.Event.Helper
 
         protected virtual void OnTriggerEnter2D(Collider2D collision)
         {
+            if (m_arrTargetTag.HasData() && !collision.CompareTag(m_arrTargetTag))
+                return;
+
             if (!IsPassCondition())
                 return;
 
@@ -49,6 +60,9 @@ namespace DSC.Event.Helper
 
         protected virtual void OnTriggerStay2D(Collider2D collision)
         {
+            if (m_arrTargetTag.HasData() && !collision.CompareTag(m_arrTargetTag))
+                return;
+
             if (!IsPassCondition())
                 return;
 
@@ -58,6 +72,9 @@ namespace DSC.Event.Helper
 
         protected virtual void OnTriggerExit2D(Collider2D collision)
         {
+            if (m_arrTargetTag.HasData() && !collision.CompareTag(m_arrTargetTag))
+                return;
+
             if (!IsPassCondition())
                 return;